requirements.txt
torchvision
torch
torchaudio
sentencepiece
transformers
numpy
pandas
scipy
accelerate
diffusers
evotorch
ipykernel
ipywidgets
treescope
# hpsv2
hpsv2x # this is hpsv2 with a bug fix.
image-reward
ml-collections
absl-py
inflect
pydantic
ipdb
open_clip_torch
jupyterlab
wandb[media]
bitsandbytes
gguf
hydra-core
beautifulsoup4
pygad